requirements.txt
loguru
pymysql
psycopg2-binary
pymilvus
faiss-cpu
numpy
